admin/nav: make analytics the default Dashboard, fold welcome into Apps
The unified analytics grid is now /dashboard. The old 'Welcome back!'
overview (DashboardStats + PopularApps + Latest Apps) moves into a
WelcomeOverview component rendered at the top of /dashboard/apps.

/dashboard/analytics now 307s back to /dashboard via next.config.js
so existing bookmarks keep working. The standalone Analytics nav
entry is removed since the dashboard root IS the analytics page.
