+++ web/admin/next.config.js
@@ -5,6 +5,12 @@ const nextConfig = {
     ignoreDuringBuilds: true,
   },
   images: { unoptimized: true },
+  async redirects() {
+    return [
+      // Analytics moved to be the default Dashboard. Keep old bookmarks alive.
+      { source: '/dashboard/analytics', destination: '/dashboard', permanent: false },
+    ];
+  },
 };
 
 module.exports = nextConfig;
